Update: rename toc to rehype for plugin structure .
